Move ui into a debug UI
